[ci]: add build step

+++ app/src/index.js
@@ -1,16 +1,21 @@
 import sequelize from './shared/database/database.js'
 import { usersRouter } from "./users/router.js"
 import express from 'express'
+import * as dotenv from 'dotenv'
+
+dotenv.config()
 
 const app = express()
-const PORT = 8000
+const PORT = process.env.PORT || 8000
 
-sequelize.sync({ force: true }).then(() => console.log('db is ready'))
+if (process.env.NODE_ENV !== 'test') {
+    sequelize.sync({ force: true }).then(() => console.log('db is ready'))
+}  
 
 app.use(express.json())
 app.use('/api/users', usersRouter)
-app.use('/up', (req, res) => { res.status(200).json(`up`)})
 
+app.use('/up', (req, res) => { res.status(200).json(`up`)})
 
 const server = app.listen(PORT, () => {
     console.log('Server running on port PORT', PORT)
